SIP-10171: fixing validation errors and style issues

+++ sip-web/src/app/modules/analyze/designer/settings/selected-fields/designer-selected-fields.component.ts
@@ -283,8 +283,13 @@ export class DesignerSelectedFieldsComponent implements OnInit, OnDestroy {
   }
 
   removeFilterFromTree(filter, index) {
-    this.analyzeService.deleteFilterFromTree(this.filters[0], filter.uuid);
     this.flattenedfilters.splice(index, 1);
+    if (filter.isAggregationFilter) {
+      index = index === 0 ? index + 1 : index;
+      this.filters.splice(index, 1);
+    } else {
+      this.analyzeService.deleteFilterFromTree(this.filters[0], filter.uuid);
+    }
     this.removeFilter.emit();
   }
 }
